+++ gridsome.config.js
@@ -1,4 +1,7 @@
 const isProd = process.env.NODE_ENV === 'production';
+const path = require('path');
+// eslint-disable-next-line import/no-extraneous-dependencies
+const glob = require('glob');
 
 module.exports = {
   siteName: 'Akli',
@@ -12,9 +15,9 @@ module.exports = {
           : 'GA_TRACKING_ID',
         debug: {
           enabled: !isProd,
-          sendHitTask: isProd
-        }
-      }
+          sendHitTask: isProd,
+        },
+      },
     },
     {
       use: '@gridsome/source-contentful',
@@ -25,14 +28,24 @@ module.exports = {
         accessToken: process.env.AKLIAISSAT_CONTENTFUL_TOKEN
           ? process.env.AKLIAISSAT_CONTENTFUL_TOKEN
           : 'CONTENTFUL_TOKEN',
-        typeName: 'Contentful'
-      }
-    }
+        typeName: 'Contentful',
+      },
+    },
   ],
-  chainWebpack: config => {
+  chainWebpack: (config) => {
     const svgRule = config.module.rule('svg');
 
     svgRule.uses.clear();
     svgRule.use('vue-svg-loader').loader('vue-svg-loader');
-  }
+  },
+  css: {
+    loaderOptions: {
+      scss: {
+        sourceMap: true,
+        includePaths: glob
+          .sync(path.join(__dirname, 'node_modules/@material'))
+          .map((dir) => path.dirname(dir)),
+      },
+    },
+  },
 };