+++ src/main.js
@@ -1,5 +1,10 @@
-import DefaultLayout from '~/layouts/Default.vue'
+import DefaultLayout from './layouts/Default.vue';
 
-export default function (Vue, { router, head, isClient }) {
-  Vue.component('Layout', DefaultLayout)
+export default function (Vue, { head }) {
+  head.link.push({
+    rel: 'stylesheet',
+    href: 'https://fonts.googleapis.com/icon?family=Material+Icons',
+  });
+
+  Vue.component('Layout', DefaultLayout);
 }