+++ CHANGELOG.md
@@ -2,8 +2,27 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.0.1] - 2019-04-14
+## [0.2.0] - 2019-10-21
+### Added
 
+- Card(s), RichText, TabBar and TopAppBar components.
+- Storybook integration.
+- ESLint.
+
+### Changed
+
+- Refactored all components to comply to Google Material Design methodologies.
+- Page GraphQL queries to sync with refactored Contentful content models.
+- Updated all unit tests.
+- Update Gridsome to v0.7.2.
+
+### Removed
+
+- Disclaimer, Icon, Notice, Banner, Bio, CookieConsent, CurriculumVitae and Social components.
+- Image assets.
+- gridsome.server.js.
+
+## [0.1.0] - 2019-04-14
 ### Added
 
 - Components.